+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support LoRa Alliance TR005 Draft 3 QR code format.
+- TLS support for application link in the Console.
 
 ### Changed